Fix code-review findings from PR #43
- serve.py: filter empty --api-key values (was silently disabling auth
  for all callers), peek the IP rate-limit bucket before running
  constant-time key comparisons, demote health/metrics polling logs to
  DEBUG, and extract a shared _run_generation generator so
  chat_completions/completions stop duplicating the
  lock/gen_start/metrics/finish_reason sequence across their streaming
  and non-streaming branches.
- optim.py: grad_cpu now matches the parameter's own dtype instead of
  being hardcoded fp32, so it doesn't silently truncate precision when
  mixed with bf16 moments under native_bf16 + CPU offload.
- config.py: check the fp4_linear/native_bf16 incompatibility before
  the generic dtype check, so train.dtype: nvfp4 + native_bf16 reports
  the actionable error instead of a misleading one.
- Extract cast_params_to_native_bf16 into model/load.py so train.py and
  load_transformer_from_checkpoint share one implementation.
- batching.py: blank lines before param_state_dtype_bytes.

+++ src/radiance/optim.py
@@ -539,7 +539,7 @@ class MuonWithAuxAdam(torch.optim.Optimizer):
                 state["exp_avg"] = _new_state_like(p, state_device)
                 state["exp_avg_sq"] = _new_state_like(p, state_device)
                 if offload:
-                    state["grad_cpu"] = torch.empty_like(p, device="cpu", dtype=torch.float32).pin_memory()
+                    state["grad_cpu"] = torch.empty_like(p, device="cpu").pin_memory()
                 state["step"] = 0
             state["step"] += 1
             if offload:
